Fix series/lists no longer language aware
