+++ openlibrary/plugins/worksearch/code.py
@@ -41,7 +41,7 @@ from openlibrary.plugins.worksearch.search import get_solr
 from openlibrary.solr.query_utils import fully_escape_query
 from openlibrary.utils.async_utils import async_bridge
 from openlibrary.utils.isbn import normalize_isbn
-from openlibrary.utils.request_context import req_context
+from openlibrary.utils.request_context import get_request_lang, req_context
 from openlibrary.utils.solr import (
     DEFAULT_PASS_TIME_ALLOWED,
     DEFAULT_SOLR_TIMEOUT_SECONDS,
@@ -206,17 +206,8 @@ def _get_readable_count(param: dict, search_response) -> int | None:
     return resp.num_found
 
 
-@public
-def get_request_lang() -> str:
-    """The request's UI language, safe to call from templates rendered on
-    either the legacy web.py server or the FastAPI server. The Templetor
-    global `get_lang()` reads `web.ctx.lang` directly, which isn't populated
-    by FastAPI — partials rendered there would AttributeError. Reading from
-    the unified `req_context` works on both. Falls back to 'en'."""
-    try:
-        return req_context.get().lang or "en"
-    except LookupError:
-        return "en"
+# Make public
+public(get_request_lang)
 
 
 async def get_solr_works_async(work_keys: set[str], fields: Iterable[str] | None = None, editions=False) -> dict[str, web.storage]:
